la-130 restructure project file organization and simplify page routing
