+++ src/components/layout/ProtectedLayout.tsx
