+++ src/page/logoutPage.tsx
@@ -1,5 +1,5 @@
 import { useEffect } from "react";
-import { useLogout } from "../../features/auth/hooks/useLogout";
+import { useLogout } from "../features/auth/hooks/useLogout";
 
 export const LogoutPage = () => {
   const logout = useLogout();
